% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuts_show.R
\name{tuts_show}
\alias{tuts_show}
\title{Show tutorials}
\usage{
tuts_show(which = NULL)
}
\arguments{
\item{which}{one specific tutorial name you'd like to access.
If this is unspecified, all tutorial names will be returned.}
}
\value{
a character vector of length equal to the length of
\code{which} if which is specified, and 10 otherwise.
}
\description{
The names of things are not easy to remember, but you need to
write the name of a tutorial to access it. This function
allows you to specify which of 10 tutorials you would like
to access.
}
\examples{

tuts_show()

tuts_show(which = 1)

}
